LayoutTests/fast/canvas/canvas-strokePath-shadow-expected.txt

Ensure correct behavior of canvas with path stroke shadow

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 20
PASS data[2] is within 2 of 0
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 20
PASS data[2] is within 2 of 0
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 20
PASS data[2] is within 2 of 0
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 20
PASS data[2] is within 2 of 0
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 0
PASS data[2] is within 2 of 0
PASS data[3] is not 255
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 0
PASS data[2] is within 2 of 0
PASS data[3] is not 255
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 0
PASS data[2] is within 2 of 0
PASS data[3] is not 255
PASS data[0] is within 2 of 255
PASS data[1] is within 2 of 0
PASS data[2] is within 2 of 0
PASS data[3] is not 255
PASS successfullyParsed is true

TEST COMPLETE
